+++ ui/src/alerts/apis/index.js
@@ -1,9 +1,10 @@
 import {proxy} from 'utils/queryUrlGenerator'
 
-export function getAlerts(source, timeRange) {
-  return proxy({
+export const getAlerts = (source, timeRange, limit) =>
+  proxy({
     source,
-    query: `SELECT host, value, level, alertName FROM alerts WHERE time >= '${timeRange.lower}' AND time <= '${timeRange.upper}' ORDER BY time desc`,
+    query: `SELECT host, value, level, alertName FROM alerts WHERE time >= '${timeRange.lower}' AND time <= '${timeRange.upper}' ORDER BY time desc ${limit
+      ? `LIMIT ${limit}`
+      : ''}`,
     db: 'chronograf',
   })
-}