+++ ui/src/alerts/components/AlertsTable.js
@@ -1,36 +1,27 @@
-import React, {PropTypes} from 'react'
+import React, {Component, PropTypes} from 'react'
 import _ from 'lodash'
 import {Link} from 'react-router'
 
-const AlertsTable = React.createClass({
-  propTypes: {
-    alerts: PropTypes.arrayOf(
-      PropTypes.shape({
-        name: PropTypes.string,
-        time: PropTypes.string,
-        value: PropTypes.string,
-        host: PropTypes.string,
-        level: PropTypes.string,
-      })
-    ),
-    source: PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
-    }).isRequired,
-  },
+class AlertsTable extends Component {
+  constructor(props) {
+    super(props)
 
-  getInitialState() {
-    return {
+    this.state = {
       searchTerm: '',
       filteredAlerts: this.props.alerts,
       sortDirection: null,
       sortKey: null,
     }
-  },
+
+    this.filterAlerts = ::this.filterAlerts
+    this.changeSort = ::this.changeSort
+    this.sortableClasses = ::this.sortableClasses
+    this.sort = ::this.sort
+  }
 
   componentWillReceiveProps(newProps) {
     this.filterAlerts(this.state.searchTerm, newProps.alerts)
-  },
+  }
 
   filterAlerts(searchTerm, newAlerts) {
     const alerts = newAlerts || this.props.alerts
@@ -47,7 +38,7 @@ const AlertsTable = React.createClass({
       )
     })
     this.setState({searchTerm, filteredAlerts})
-  },
+  }
 
   changeSort(key) {
     // if we're using the key, reverse order; otherwise, set it with ascending
@@ -59,7 +50,17 @@ const AlertsTable = React.createClass({
     } else {
       this.setState({sortKey: key, sortDirection: 'asc'})
     }
-  },
+  }
+
+  sortableClasses(key) {
+    if (this.state.sortKey === key) {
+      if (this.state.sortDirection === 'asc') {
+        return 'sortable-header sorting-ascending'
+      }
+      return 'sortable-header sorting-descending'
+    }
+    return 'sortable-header'
+  }
 
   sort(alerts, key, direction) {
     switch (direction) {
@@ -70,125 +71,167 @@ const AlertsTable = React.createClass({
       default:
         return alerts
     }
-  },
+  }
 
-  render() {
-    const {id} = this.props.source
+  renderTable() {
+    const {source: {id}} = this.props
     const alerts = this.sort(
       this.state.filteredAlerts,
       this.state.sortKey,
       this.state.sortDirection
     )
-    return (
-      <div className="panel panel-minimal">
-        <div className="panel-heading u-flex u-ai-center u-jc-space-between">
-          <h2 className="panel-title">{this.props.alerts.length} Alerts</h2>
-          {this.props.alerts.length
-            ? <SearchBar onSearch={this.filterAlerts} />
+    return this.props.alerts.length
+      ? <table className="table v-center table-highlight">
+          <thead>
+            <tr>
+              <th
+                onClick={() => this.changeSort('name')}
+                className={this.sortableClasses('name')}
+              >
+                Name
+              </th>
+              <th
+                onClick={() => this.changeSort('level')}
+                className={this.sortableClasses('level')}
+              >
+                Level
+              </th>
+              <th
+                onClick={() => this.changeSort('time')}
+                className={this.sortableClasses('time')}
+              >
+                Time
+              </th>
+              <th
+                onClick={() => this.changeSort('host')}
+                className={this.sortableClasses('host')}
+              >
+                Host
+              </th>
+              <th
+                onClick={() => this.changeSort('value')}
+                className={this.sortableClasses('value')}
+              >
+                Value
+              </th>
+            </tr>
+          </thead>
+          <tbody>
+            {alerts.map(({name, level, time, host, value}) => {
+              return (
+                <tr key={`${name}-${level}-${time}-${host}-${value}`}>
+                  <td className="monotype">{name}</td>
+                  <td className={`monotype alert-level-${level.toLowerCase()}`}>
+                    {level}
+                  </td>
+                  <td className="monotype">
+                    {new Date(Number(time)).toISOString()}
+                  </td>
+                  <td className="monotype">
+                    <Link to={`/sources/${id}/hosts/${host}`}>
+                      {host}
+                    </Link>
+                  </td>
+                  <td className="monotype">{value}</td>
+                </tr>
+              )
+            })}
+          </tbody>
+        </table>
+      : this.renderTableEmpty()
+  }
+
+  renderTableEmpty() {
+    const {source: {id}, shouldNotBeFilterable} = this.props
+
+    return shouldNotBeFilterable
+      ? <div className="graph-empty">
+          <p>
+            Learn how to configure your first <strong>Rule</strong> in<br />
+            the <em>Getting Started</em> guide
+          </p>
+        </div>
+      : <div className="generic-empty-state">
+          <h4 className="no-user-select">
+            There are no Alerts to display
+          </h4>
+          <br />
+          <h6 className="no-user-select">
+            Try changing the Time Range or
+            <Link
+              style={{marginLeft: '10px'}}
+              to={`/sources/${id}/alert-rules/new`}
+              className="btn btn-primary btn-sm"
+            >
+              Create an Alert Rule
+            </Link>
+          </h6>
+        </div>
+  }
+
+  render() {
+    const {
+      shouldNotBeFilterable,
+      limit,
+      onGetMoreAlerts,
+      isAlertsMaxedOut,
+      alertsCount,
+    } = this.props
+
+    return shouldNotBeFilterable
+      ? <div className="alerts-widget">
+          {this.renderTable()}
+          {limit && alertsCount
+            ? <button
+                className="btn btn-sm btn-default btn-block"
+                onClick={onGetMoreAlerts}
+                disabled={isAlertsMaxedOut}
+                style={{marginBottom: '20px'}}
+              >
+                {isAlertsMaxedOut
+                  ? `All ${alertsCount} Alerts displayed`
+                  : 'Load next 30 Alerts'}
+              </button>
             : null}
         </div>
-        <div className="panel-body">
-          {this.props.alerts.length
-            ? <table className="table v-center table-highlight">
-                <thead>
-                  <tr>
-                    <th
-                      onClick={() => this.changeSort('name')}
-                      className="sortable-header"
-                    >
-                      Name
-                    </th>
-                    <th
-                      onClick={() => this.changeSort('level')}
-                      className="sortable-header"
-                    >
-                      Level
-                    </th>
-                    <th
-                      onClick={() => this.changeSort('time')}
-                      className="sortable-header"
-                    >
-                      Time
-                    </th>
-                    <th
-                      onClick={() => this.changeSort('host')}
-                      className="sortable-header"
-                    >
-                      Host
-                    </th>
-                    <th
-                      onClick={() => this.changeSort('value')}
-                      className="sortable-header"
-                    >
-                      Value
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {alerts.map(({name, level, time, host, value}) => {
-                    return (
-                      <tr key={`${name}-${level}-${time}-${host}-${value}`}>
-                        <td className="monotype">{name}</td>
-                        <td
-                          className={`monotype alert-level-${level.toLowerCase()}`}
-                        >
-                          {level}
-                        </td>
-                        <td className="monotype">
-                          {new Date(Number(time)).toISOString()}
-                        </td>
-                        <td className="monotype">
-                          <Link to={`/sources/${id}/hosts/${host}`}>
-                            {host}
-                          </Link>
-                        </td>
-                        <td className="monotype">{value}</td>
-                      </tr>
-                    )
-                  })}
-                </tbody>
-              </table>
-            : <div className="generic-empty-state">
-                <h5 className="no-user-select">
-                  Alerts appear here when you have Rules
-                </h5>
-                <br />
-                <Link
-                  to={`/sources/${id}/alert-rules/new`}
-                  className="btn btn-primary"
-                >
-                  Create a Rule
-                </Link>
-              </div>}
+      : <div className="panel panel-minimal">
+          <div className="panel-heading u-flex u-ai-center u-jc-space-between">
+            <h2 className="panel-title">{this.props.alerts.length} Alerts</h2>
+            {this.props.alerts.length
+              ? <SearchBar onSearch={this.filterAlerts} />
+              : null}
+          </div>
+          <div className="panel-body">
+            {this.renderTable()}
+          </div>
         </div>
-      </div>
-    )
-  },
-})
+  }
+}
 
-const SearchBar = React.createClass({
-  propTypes: {
-    onSearch: PropTypes.func.isRequired,
-  },
+class SearchBar extends Component {
+  constructor(props) {
+    super(props)
 
-  getInitialState() {
-    return {
+    this.state = {
       searchTerm: '',
     }
-  },
+
+    this.handleSearch = ::this.handleSearch
+    this.handleChange = ::this.handleChange
+  }
 
   componentWillMount() {
     const waitPeriod = 300
     this.handleSearch = _.debounce(this.handleSearch, waitPeriod)
-  },
+  }
 
   handleSearch() {
     this.props.onSearch(this.state.searchTerm)
-  },
+  }
 
   handleChange(e) {
     this.setState({searchTerm: e.target.value}, this.handleSearch)
-  },
+  }
 
   render() {
     return (
@@ -205,7 +248,34 @@ const SearchBar = React.createClass({
         </div>
       </div>
     )
-  },
-})
+  }
+}
+
+const {arrayOf, bool, func, number, shape, string} = PropTypes
+
+AlertsTable.propTypes = {
+  alerts: arrayOf(
+    shape({
+      name: string,
+      time: string,
+      value: string,
+      host: string,
+      level: string,
+    })
+  ),
+  source: shape({
+    id: string.isRequired,
+    name: string.isRequired,
+  }).isRequired,
+  shouldNotBeFilterable: bool,
+  limit: number,
+  onGetMoreAlerts: func,
+  isAlertsMaxedOut: bool,
+  alertsCount: number,
+}
+
+SearchBar.propTypes = {
+  onSearch: func.isRequired,
+}
 
 export default AlertsTable