+++ ui/src/alerts/containers/AlertsApp.js
@@ -6,15 +6,24 @@ import NoKapacitorError from 'shared/components/NoKapacitorError'
 import CustomTimeRangeDropdown from 'shared/components/CustomTimeRangeDropdown'
 import FancyScrollbar from 'shared/components/FancyScrollbar'
 
-import {getAlerts} from '../apis'
+import {getAlerts} from 'src/alerts/apis'
 import AJAX from 'utils/ajax'
 
 import _ from 'lodash'
 import moment from 'moment'
 
+import timeRanges from 'hson!shared/data/timeRanges.hson'
+
 class AlertsApp extends Component {
   constructor(props) {
     super(props)
+
+    const lowerInSec = props.timeRange
+      ? timeRanges.find(tr => tr.lower === props.timeRange.lower).seconds
+      : undefined
+
+    const oneDayInSec = 86400
+
     this.state = {
       loading: true,
       hasKapacitor: false,
@@ -22,12 +31,16 @@ class AlertsApp extends Component {
       isTimeOpen: false,
       timeRange: {
         upper: moment().format(),
-        lower: moment().subtract(1, 'd').format(),
+        lower: moment().subtract(lowerInSec || oneDayInSec, 'seconds').format(),
       },
+      limit: props.limit || 0, // only used if AlertsApp receives a limit prop
+      limitMultiplier: 1, // only used if AlertsApp receives a limit prop
+      isAlertsMaxedOut: false, // only used if AlertsApp receives a limit prop
     }
 
     this.fetchAlerts = ::this.fetchAlerts
     this.renderSubComponents = ::this.renderSubComponents
+    this.handleGetMoreAlerts = ::this.handleGetMoreAlerts
     this.handleToggleTime = ::this.handleToggleTime
     this.handleCloseTime = ::this.handleCloseTime
     this.handleApplyTime = ::this.handleApplyTime
@@ -59,7 +72,8 @@ class AlertsApp extends Component {
   fetchAlerts() {
     getAlerts(
       this.props.source.links.proxy,
-      this.state.timeRange
+      this.state.timeRange,
+      this.state.limit * this.state.limitMultiplier
     ).then(resp => {
       const results = []
 
@@ -90,23 +104,39 @@ class AlertsApp extends Component {
           name: `${s[nameIndex]}`,
         })
       })
-      this.setState({loading: false, alerts: results})
+
+      // TODO: factor these setStates out to make a pure function and implement true limit & offset
+      this.setState({
+        loading: false,
+        alerts: results,
+        // this.state.alerts.length === results.length ||
+        isAlertsMaxedOut:
+          results.length !== this.props.limit * this.state.limitMultiplier,
+      })
+    })
+  }
+
+  handleGetMoreAlerts() {
+    this.setState({limitMultiplier: this.state.limitMultiplier + 1}, () => {
+      this.fetchAlerts(this.state.limitMultiplier)
     })
   }
 
   renderSubComponents() {
-    const {source} = this.props
-    return (
-      <div className="container-fluid">
-        <div className="row">
-          <div className="col-md-12">
-            {this.state.hasKapacitor
-              ? <AlertsTable source={source} alerts={this.state.alerts} />
-              : <NoKapacitorError source={source} />}
-          </div>
-        </div>
-      </div>
-    )
+    const {source, isWidget, limit} = this.props
+    const {isAlertsMaxedOut, alerts} = this.state
+
+    return this.state.hasKapacitor
+      ? <AlertsTable
+          source={source}
+          alerts={this.state.alerts}
+          shouldNotBeFilterable={isWidget}
+          limit={limit}
+          onGetMoreAlerts={this.handleGetMoreAlerts}
+          isAlertsMaxedOut={isAlertsMaxedOut}
+          alertsCount={alerts.length}
+        />
+      : <NoKapacitorError source={source} />
   }
 
   handleToggleTime() {
@@ -122,43 +152,51 @@ class AlertsApp extends Component {
   }
 
   render() {
-    const {source} = this.props
+    const {isWidget, source} = this.props
     const {loading, timeRange} = this.state
 
     if (loading || !source) {
       return <div className="page-spinner" />
     }
 
-    return (
-      <div className="page">
-        <div className="page-header">
-          <div className="page-header__container">
-            <div className="page-header__left">
-              <h1 className="page-header__title">
-                Alert History
-              </h1>
-            </div>
-            <div className="page-header__right">
-              <SourceIndicator sourceName={source.name} />
-              <CustomTimeRangeDropdown
-                isVisible={this.state.isTimeOpen}
-                onToggle={this.handleToggleTime}
-                onClose={this.handleCloseTime}
-                onApplyTimeRange={this.handleApplyTime}
-                timeRange={timeRange}
-              />
-            </div>
-          </div>
-        </div>
-        <FancyScrollbar className="page-contents">
+    return isWidget
+      ? <FancyScrollbar autoHide={false}>
           {this.renderSubComponents()}
         </FancyScrollbar>
-      </div>
-    )
+      : <div className="page">
+          <div className="page-header">
+            <div className="page-header__container">
+              <div className="page-header__left">
+                <h1 className="page-header__title">
+                  Alert History
+                </h1>
+              </div>
+              <div className="page-header__right">
+                <SourceIndicator sourceName={source.name} />
+                <CustomTimeRangeDropdown
+                  isVisible={this.state.isTimeOpen}
+                  onToggle={this.handleToggleTime}
+                  onClose={this.handleCloseTime}
+                  onApplyTimeRange={this.handleApplyTime}
+                  timeRange={timeRange}
+                />
+              </div>
+            </div>
+          </div>
+          <FancyScrollbar className="page-contents">
+            <div className="container-fluid">
+              <div className="row">
+                <div className="col-md-12">
+                  {this.renderSubComponents()}
+                </div>
+              </div>
+            </div>
+          </FancyScrollbar>
+        </div>
   }
 }
 
-const {func, shape, string} = PropTypes
+const {bool, number, oneOfType, shape, string} = PropTypes
 
 AlertsApp.propTypes = {
   source: shape({
@@ -169,7 +207,12 @@ AlertsApp.propTypes = {
       proxy: string.isRequired,
     }).isRequired,
   }),
-  addFlashMessage: func,
+  timeRange: shape({
+    lower: string.isRequired,
+    upper: oneOfType([shape(), string]),
+  }),
+  isWidget: bool,
+  limit: number,
 }
 
 export default AlertsApp