+++ ui/src/shared/reducers/index.js
@@ -1,6 +1,7 @@
 import app from './app'
 import auth from './auth'
 import errors from './errors'
+import links from './links'
 import notifications from './notifications'
 import sources from './sources'
 
@@ -8,6 +9,7 @@ export default {
   app,
   auth,
   errors,
+  links,
   notifications,
   sources,
 }