+++ ui/src/status/reducers/index.js
@@ -1,0 +1,7 @@
+import statusUI from './ui'
+import JSONFeed from './JSONFeed'
+
+export default {
+  statusUI,
+  JSONFeed,
+}